requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
deep-translator==1.11.4
Django==5.1.3
django-deep-translator==1.5.1
django-filebrowser==4.0.3
django-grappelli==3.0.9
django-rosetta==0.10.1
django-tinymce==4.1.0
djlint==1.36.1
EditorConfig==0.12.4
gunicorn==23.0.0
idna==3.10
jsbeautifier==1.15.1
json5==0.9.28
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
polib==1.2.0
psycopg2-binary==2.9.10
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
tqdm==4.67.0
tzdata==2025.1
urllib3==2.3.0
